add cookies for authentication

+++ src/routes/home.js
@@ -1,39 +1,64 @@
 var express = require('express');
 var router = express.Router();
 const db = require('../helpers/db');
+const session = require('../helpers/session')
 
 
-router.post("/home", (req, res) => {
-  task = req.body.task
-  s = "ongoing"
-  let query = "INSERT INTO Todo (task, status, user_id) VALUES ('" + task + "', '" + s + "', '" + req.session.user_id + "');"
+router.post("/home", async(req, res) => {
 
-  db.connection.query(query, function (err, result) {
+  // check if user is already logged in
+  var user = await session.checkSessionId(req.cookies.sessionid);
+  if (user == undefined) {  
+    res.redirect('/');
+  } else {
+    task = req.body.task
+    s = "ongoing"
+    let query = "INSERT INTO Todo (task, status, user_id) VALUES ('" + task + "', '" + s + "', '" + req.session.user_id + "');"
+  
+    db.connection.query(query, function (err, result) {
+        if (err) throw err;
+        res.redirect('/home')
+    });
+  }
+
+})
+
+router.get("/home/:id", async (req, res) =>{
+
+  // check if user is already logged in
+  var user = await session.checkSessionId(req.cookies.sessionid);
+  if (user == undefined) {  
+    res.redirect('/');
+  } else {
+    let query = "DELETE FROM Todo WHERE task_id='" + req.params.id + "';"
+    db.connection.query(query, function (err, result) {
       if (err) throw err;
       res.redirect('/home')
-  });
-
-})
-
-router.get("/home/:id", (req, res) =>{
-  let query = "DELETE FROM Todo WHERE task_id='" + req.params.id + "';"
-  db.connection.query(query, function (err, result) {
-    if (err) throw err;
-    res.redirect('/home')
-  });
-})
-
-router.get('/home', function(request, response, next) {
-  let query = "SELECT * FROM Todo WHERE user_id='" + request.session.user_id + "';"
-  let items = []
-  db.connection.query(query, function (err, result) {
-    if (err) throw err;
-    items = result
-    response.render('index', {
-      loggedIn: request.session.loggedin,
-      items: items
     });
-  });
+  }
+
+})
+
+router.get('/home', async function(request, response, next) {
+
+
+  // check if user is already logged in
+  var user = await session.checkSessionId(request.cookies.sessionid);
+  if (user == undefined) {  
+    response.redirect('/');
+  } else {
+    let query = "SELECT * FROM Todo WHERE user_id='" + request.session.user_id + "';"
+    let items = []
+    db.connection.query(query, function (err, result) {
+      if (err) throw err;
+      items = result
+      response.render('index', {
+        loggedIn: true,
+        items: items
+      });
+    });
+  }
+
 });
